Melhorado métodos de extensão.
