+++ src/SME.SGP.Aula.Worker/Startup.cs
@@ -29,8 +29,8 @@ namespace SME.SGP.Aula.Worker
             registrarDependencias.RegistrarCasoDeUsoAulaRabbitSgp(services);
 
             services.AddHostedService<WorkerRabbitAula>();
-            services.AddHealthChecksSgp().Builder();
-            services.AddHealthChecksUiSgp().Builder();            
+            services.AddHealthChecks();
+            services.AddHealthChecksUiSgp();            
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
