+++ src/SME.SGP.Fechamento.Worker/Startup.cs
@@ -29,8 +29,8 @@ namespace SME.SGP.Fechamento.Worker
             registrarDependencias.RegistrarCasoDeUsoFechamentoRabbitSgp(services);
 
             services.AddHostedService<WorkerRabbitFechamento>();
-            services.AddHealthChecksSgp().Builder();
-            services.AddHealthChecksUiSgp().Builder();
+            services.AddHealthChecks();
+            services.AddHealthChecksUiSgp();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
