+++ src/SME.SGP.Frequencia.Worker/Startup.cs
@@ -29,8 +29,8 @@ namespace SME.SGP.Frequencia.Worker
             registrarDependencias.RegistrarCasoDeUsoFrequenciaRabbitSgp(services);
 
             services.AddHostedService<WorkerRabbitFrequencia>();
-            services.AddHealthChecksSgp().Builder();
-            services.AddHealthChecksUiSgp().Builder();
+            services.AddHealthChecks();
+            services.AddHealthChecksUiSgp();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
