+++ src/SME.SGP.Pendencias.Worker/Startup.cs
@@ -29,8 +29,8 @@ namespace SME.SGP.Pendencias.Worker
             registrarDependencias.RegistrarCasoDeUsoPendenciasRabbitSgp(services);
 
             services.AddHostedService<WorkerRabbitPendencias>();
-            services.AddHealthChecksSgp().Builder();
-            services.AddHealthChecksUiSgp().Builder();
+            services.AddHealthChecks();
+            services.AddHealthChecksUiSgp();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
