+++ src/SME.SGP.Worker.Rabbbit/Startup.cs
@@ -32,8 +32,8 @@ namespace SME.SGP.Worker.Rabbbit
             registrarDependencias.RegistrarCasoDeUsoRabbitSgp(services);
 
             services.AddHostedService<WorkerRabbitMQ>();
-            services.AddHealthChecksSgp().Builder();
-            services.AddHealthChecksUiSgp().Builder();
+            services.AddHealthChecks();
+            services.AddHealthChecksUiSgp();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
